requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
JPype1==1.4.1
h2==2.2.220
sqlalchemy==2.0.23
pandas==2.0.3
reportlab==4.0.4
openpyxl==3.1.2
